Converting java objects to native
We decided it would be better to store the objects natively
to cut down on overhead associated with crossing the c/java
barrier.  Instead, we will store the pointers in the form
of longs on the java side, and allow the details of the
objects to reside on the c side of the code.

Signed-off-by: Nathaniel Graham <ngraham@lanl.gov>