May be working.
I have fixed all the crashes.  I commented out the exit
call in fabric.c on line 33.  On my machine this was being
called.  I am not sure if it is because my machine does
not have the necessary things in place to actually get
an info object back, but I will try to test this on a
machine I know it works on.

Signed-off-by: Nathaniel Graham <ngraham@lanl.gov>