Version object added and corrections to fi_getinfo bindings
This commit includes the addition of a Version object used
to store the version of the provider.  It also includes
fixes to the fi_getinfo bindings based on feedback from
Howard.  In addition, this commit includes a fair ammount
of code restructuring and coding style level changes from
things I found while relearning the code after having not
touched it in so long a time.

Signed-off-by: Nathaniel Graham <ngraham@lanl.gov>