Handle null inputs for getInfo
In the libfabric documentation it says the user is allowed
to use NULL for node, service, and hints.  I have made
modifications to the Java bindings to allow the use of null
for those variables.  This also cleared up an issue where
I was not getting all of the potential providers shown by
the fi_info executable in the bin of the libfabric install.
Also renamed the test that getInfo is in to mirror the move
from Fabric.java to LibFabric.java.

Signed-off-by: Nathaniel Graham <ngraham@lanl.gov>