Updated getInfo and added FI_SOURCE
Updated getInfo to include a method call that does not require
and info object.  This is a much cleaner way to handle the way
you can pass null in the C version of libfabric.

Also added the FI_SOURCE constant which is needed in the pingpong
test.

This commit also includes the most recent changes to the pingpong
test to create a more recent backup copy.

Signed-off-by: Nathaniel Graham <ngraham@lanl.gov>