More PingPong work
This commit includes more changes to the pingpong test and
changes to the java bindings to support these changes.
Added an overridden equals method to FIDescriptor which
checks whether the handles are the same (the underlying
C structs are the same).  Also added a few more constants
that were needed by the test.

Note: Attempting to keep the changes made to the pingpong
test separate from updates made to the java bindings is
creating unnecessary complications.  In the future, I will
simply update the pingpong test when I push needed changes
to the Java bindings.

Signed-off-by: Nathaniel Graham <ngraham@lanl.gov>