Change inheritance structure for endpoints
I originally had it so passive endpoints inherited from
endpoints because my initial impression was that passive
endpoints are a subclass of normal endpoints.  This was
incorrect.  Passive endpoints and endpoints are distinctly
different objects, but they still share some methods such
as get and set name.  To this end, I have created a new
class EndPointSharedOps (needs a better name) to hold said
shared methods, and both passive and normal endpoints
inherit from it.

This commit also includes the most recent changes for
porting the pingpong test to java.

Signed-off-by: Nathaniel Graham <ngraham@lanl.gov>