Updated the free info method to use libfabric's freeinfo
Updated the free info method in Info.java to use libfabric's
freeinfo method.  Note, currently the constructor that takes
arguments for all of the info attributes does not work correctly.
Some of the fields in Info need to be alloced, because the
libfabric freeinfo method will try to free them, causing a
crash.

This commit also includes the updates to the pingpong test
which triggered these changes.

Signed-off-by: Nathaniel Graham <ngraham@lanl.gov>