Added the mrRegister method to Domain
Added the ability to register a memory region to the Domain
object.  I also fixed an issue in DomainAttr where there
were numerous ints which should have been longs. Added
another constant that was needed.

This commit also includes changes to the ping pong test
which prompted the changes to the java bindings.

Signed-off-by: Nathaniel Graham <ngraham@lanl.gov>