Added overloaded send and getDesc for MemoryRegion
Added another overloaded send method to EndPoint and
a getDesc method for MemoryRegion.  The getDesc method
is a quick and dirty fix that should be cleaned up if
these bindings are ever released for public use.

This commit also includes changes to the ping pong
test which prompted these changes.

Signed-off-by: Nathaniel Graham <ngraham@lanl.gov>