Added the Message object and the sendMessage method
I added the Message object and the sendMessage method
in EndPoint.

I also removed one of the WaitObj values
that was removed in the main libfabric project.

The method to set the values stored in Constant was
not setting the values correctly.  I fixed that as
well.

This commit also includes the changes to the PingPongTest
that prompted the changes to the Java bindings.

Signed-off-by: Nathaniel Graham <ngraham@lanl.gov>