Load all classes on LibFabric init
This commit includes many changes that were made
while debugging PingPongTest.java, but the primary
change is that all classes are now being loaded
when libfabric is initted.  The EndPoint class
was not getting loaded by the JVM in the test. If
the class was not loaded before the sockets were
used to send some control information, the program
would crash because the JVM could no longer find
it.  To solve this issue, and any related issues,
I manually load all of my classes when the init
method is called.

Signed-off-by: Nathaniel Graham <ngraham@lanl.gov>