Remove debugging code
Removed a lot of code that was used to debug some problems I
was having in PingPongTest.java.  In the process of cleaning
up the test, I discovered that the ep recv was throwing an
error that was simply being caught and ignored.  A future
commit will fix that issue.

Signed-off-by: Nathaniel Graham <ngraham@lanl.gov>