PingPongTest completed
This commit includes the final touches to getting PingPongTest.java
working.  There may still be an issue with the test as I am getting far
better performance out of the Java bindings than from the C itself, but
it may be a result of just running the test on my local machine.  Will
be running the test on a cluster to see how performance results are
then.

Signed-off-by: Nathaniel Graham <ngraham@lanl.gov>